feat: expenses and files

+++ app/Http/Controllers/Api/ExpenseController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers\Api;
 
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Carbon\Carbon;
+use App\Models\Permission;
 use App\Models\Expense;
 use App\Http\Controllers\Controller;
 
@@ -11,6 +13,8 @@ class ExpenseController extends Controller
 {
     public function index(Request $request)
     {
+        Gate::authorize(Permission::VIEW_EXPENSES);
+
         $location = $request->user()->location;
         $perPage = $request->input('rows', 10);
         $expenses = Expense::with(['expenseCategory', 'creator', 'location'])
@@ -23,6 +27,8 @@ class ExpenseController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize(Permission::CREATE_EXPENSES);
+
         $request->validate([
             'amount' => 'required|numeric',
             'category_id' => 'required|exists:expense_categories,id',
@@ -46,6 +52,8 @@ class ExpenseController extends Controller
 
     public function update(Request $request, Expense $expense)
     {
+        Gate::authorize(Permission::UPDATE_EXPENSES);
+
         $request->validate([
             'amount' => 'required|numeric',
             'category_id' => 'required|exists:expense_categories,id',
@@ -67,6 +75,8 @@ class ExpenseController extends Controller
 
     public function destroy(Expense $expense)
     {
+        Gate::authorize(Permission::DELETE_EXPENSES);
+
         $expense->delete();
 
         return response()->json([
